timeline: Fix timeline layer controls expanding the parent widget
This wasn't needed in gtk2 because all widgets could shrink. Scrolled window +
viewport is supposed to provide the same shrinking logic for gtk3. That would
only work without scrollbar policy NEVER, so we need this little scrollbar hide
hack to get what we need.

Setting "shrink" on the paned doesn't fix this either, because we still
wouldn't be able to shrink the controls at all.